requirements.txt
django==4.2
social-auth-app-django==5.4.0
markdown==3.5.1
django-parler==2.3
